requirements.txt

krakenex
pandas
numpy
ta
schedule
requests
python-dotenv
ccxt